+++ libraries/Microsoft.Bot.Connector/Microsoft.Bot.Connector.csproj
@@ -26,19 +26,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="node_modules\**" />
-    <Compile Remove="Skills\**" />
-    <EmbeddedResource Remove="node_modules\**" />
-    <EmbeddedResource Remove="Skills\**" />
-    <None Remove="node_modules\**" />
-    <None Remove="Skills\**" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Remove="generateClient.cmd" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Logging" Version="2.1.0" />
     <PackageReference Include="Microsoft.Extensions.Http" Version="2.1.0" />
     <PackageReference Include="Microsoft.IdentityModel.Clients.ActiveDirectory" Version="5.2.4" />
