+++ tests/Auth/bot-authentication/AuthenticationBot.csproj
@@ -20,7 +20,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <None Update="main.dialog">
+    <None Update="test.dialog">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
   </ItemGroup>
