+++ tests/Auth/bot-authentication/appsettings.json
@@ -1,5 +1,6 @@
 {
   "MicrosoftAppId": "",
   "MicrosoftAppPassword": "",
-  "ConnectionName": ""
+  "ConnectionName": "",
+  "defaultRootDialog": "test.dialog"
 }
