+++ tests/Auth/bot-authentication/test.dialog
